Refactor HPC schedulers into unified strategy module
